requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
streamlit
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo_search==5.3.1b1
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
gradio
langchain==0.2.16
langchain-core>=0.2.38
pydantic>=2.7.4
langchain_nvidia_ai_endpoints
crewai
crewai_tools
load_dotenv
langchain-huggingface
yt_dlp==2023.11.14 
youtube-transcript-api==0.6.1


